requirements.txt
python-telegram-bot==21.10
aiofiles==24.1.0
pydantic==2.10.6
python-dotenv==1.0.1
yfinance==0.2.54
pandas==2.2.3
matplotlib==3.10.0
streamlit==1.42.0
